# quarkus-jpa-start/target/jib/META-INF/quarkus-generated-openapi-doc.JSON
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "quarkus-jpa-start API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "paths" : {
...
    "/api/api/locations/add" : {
      "post" : {
        "tags" : [ "Location Resource" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Locations"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/api/locations/projects/{id}" : {
      "get" : {
        "tags" : [ "Location Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Projects"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/api/locations/{id}" : {
      "get" : {
        "tags" : [ "Location Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Locations"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/api/projects/{id}" : {
      "get" : {
        "tags" : [ "Projects Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Projects"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Customer" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "tel" : {
            "type" : "string"
          }
        }
      },
...
      "Locations" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "latitude" : {
            "format" : "float",
            "type" : "number"
          },
          "longitude" : {
            "format" : "float",
            "type" : "number"
          },
          "info" : {
            "type" : "string"
          }
        }
      },
      "Projects" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "customer" : {
            "$ref" : "#/components/schemas/Customer"
          },
          "description" : {
            "type" : "string"
          },
          "customerid" : {
            "$ref" : "#/components/schemas/Customer"
          }
        }
      },
...
    }
  }
}